[refactor] 임의로 추가한 이미지 변경
- 임의로 추가한 이미지 제거 및 새로운 이미지 적용
